requirements.txt
labelme==6.0.0
matplotlib==3.10.8
numpy==2.4.4
Pillow==12.2.0
scikit_learn==1.8.0
thop==0.1.1.post2209072238
torchsummary==1.5.1
tqdm==4.67.3
wandb==0.25.1
opencv-python==4.13.0